// generated-api/models/category.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Solutions API
 * REST API for technical solutions organized by categories and subcategories
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Category
 */
export interface Category {
    /**
     * 
     * @type {any}
     * @memberof Category
     */
    'id'?: any;
    /**
     * 
     * @type {any}
     * @memberof Category
     */
    'name'?: any;
    /**
     * 
     * @type {any}
     * @memberof Category
     */
    'slug'?: any;
}