// generated-api/models/subcategory.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Solutions API
 * REST API for technical solutions organized by categories and subcategories
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Subcategory
 */
export interface Subcategory {
    /**
     * 
     * @type {any}
     * @memberof Subcategory
     */
    'id'?: any;
    /**
     * 
     * @type {any}
     * @memberof Subcategory
     */
    'name'?: any;
    /**
     * 
     * @type {any}
     * @memberof Subcategory
     */
    'slug'?: any;
    /**
     * 
     * @type {any}
     * @memberof Subcategory
     */
    'category_id'?: any;
}